refactor: move downstack dependencies api call to api module
