+++ src/lib/api/common_params.ts
@@ -1,0 +1,5 @@
+export type TRepoParams = {
+  authToken: string;
+  repoName: string;
+  repoOwner: string;
+};
